+++ src/moonlink/src/storage/mooncake_table/table_snapshot.rs
@@ -1,7 +1,11 @@
+use crate::storage::iceberg::puffin_utils::PuffinBlobRef;
 /// Items needed for iceberg snapshot.
 use crate::storage::index::FileIndex as MooncakeFileIndex;
 use crate::storage::mooncake_table::delete_vector::BatchDeletionVector;
 use crate::storage::storage_utils::MooncakeDataFileRef;
+use crate::storage::TableManager;
+
+use std::collections::HashMap;
 
 #[derive(Debug)]
 pub struct IcebergSnapshotPayload {
@@ -15,4 +19,16 @@ pub struct IcebergSnapshotPayload {
     ///
     /// TODO(hjiang): It's ok to take only new moooncake file index, instead of all file indices.
     pub(crate) file_indices: Vec<MooncakeFileIndex>,
+}
+
+/// Return type of async iceberg snapshot creation.
+pub(crate) struct IcebergSnapshotResult {
+    /// Table manager is (1) not `Sync` safe; (2) only used at iceberg snapshot creation, so we `move` it around every snapshot.
+    pub(crate) table_manager: Box<dyn TableManager>,
+    /// Iceberg flush LSN.
+    pub(crate) flush_lsn: u64,
+    /// Persisted data files.
+    pub(crate) new_data_files: Vec<MooncakeDataFileRef>,
+    /// Persisted puffin blob reference.
+    pub(crate) puffin_blob_ref: HashMap<MooncakeDataFileRef, PuffinBlobRef>,
 }
